refactor: rename variables name and improving code indentation
